created a new create trip page and made the button navigate to it
