+++ client/src/App.tsx
@@ -15,6 +15,7 @@ import CreateConference from "@/pages/CreateConference";
 import EditConference from "@/pages/EditConference";
 import CreateWorkshop from "@/pages/CreateWorkshop";
 import CreateTrip from "@/pages/CreateTrip";
+import CreateTripForm from "@/pages/CreateTripForm";
 import EditTrip from "@/pages/EditTrip";
 import CreateBazaar from "@/pages/CreateBazaar";
 import VendorDashboard from "@/pages/VendorDashboard";
@@ -111,9 +112,14 @@ function Router() {
           <EditWorkshop />
         </ProtectedRoute>
       </Route>
-      <Route path="/create/trip">
+      <Route path="/trips">
         <ProtectedRoute allowedRoles={["events_office"]}>
           <CreateTrip />
+        </ProtectedRoute>
+      </Route>
+      <Route path="/events-office/create/trip">
+        <ProtectedRoute allowedRoles={["events_office"]}>
+          <CreateTripForm />
         </ProtectedRoute>
       </Route>
       <Route path="/events-office/events/trip/edit/:id">
