+++ client/src/components/EventsOfficeHeader.tsx
@@ -89,8 +89,8 @@ export default function EventsOfficeHeader() {
           <Button
             variant="ghost"
             size="sm"
-            className={`gap-2 ${location === "/create/trip" ? "underline decoration-primary decoration-2" : ""}`}
-            onClick={() => setLocation("/create/trip")}
+            className={`gap-2 ${location === "/trips" ? "underline decoration-primary decoration-2" : ""}`}
+            onClick={() => setLocation("/trips")}
             data-testid="button-nav-trips"
           >
             <Plane className="h-4 w-4" />
